Added lazy loading for images and improve page load
This commit adds lazy loading for images in the App component to improve page load performance. It also updates the App component to use the useLocation hook from react-router-dom. Additionally, the Navbar component is updated to include a link to the Typing page. The ProtectedLayout component is removed as it is no longer needed.