feat: Add calculateAccuracy function to Results component
This commit adds the calculateAccuracy function to the Results component in order to calculate the accuracy of the user's typing. The function takes the total number of characters and the number of mistakes as parameters and returns the accuracy percentage. This function is used to display the accuracy in the Results component.